优化训练脚本

+++ script/run/run_sft.sh
@@ -1,6 +1,9 @@
 #!/bin/bash
 
+nproc_per_node=2
+
 target=/data0/maqi/KGLQA-model/train_args/quality/option1-1-race-2-quality.json
+#target=/data0/maqi/KGLQA-model/train_args/cclue/cclue_ft.json
 
 log_file=train.log
 # 读取json文件中 output_dir 的值
@@ -9,13 +12,21 @@ output_dir=$(jq -r '.output_dir' ${target})
 if [ ! -d "${output_dir}" ];then
     mkdir -p "${output_dir}"
 else
-  # 删除目录下所有文件
-    # shellcheck disable=SC2115
-    rm -rf "${output_dir}"/*
+  # 询问是否删除 y删除 n不删除
+  # shellcheck disable=SC2162
+  read -p "是否删除 ${output_dir} 目录下所有文件？(y/n)" input
+  if [ "${input}" = "y" ]; then
+      # shellcheck disable=SC2115
+      echo "已删除 ${output_dir} 目录下所有文件"
+      # shellcheck disable=SC2115
+      rm -rf "${output_dir}"/*
+  else
+      echo "保留 ${output_dir} 目录下所有文件"
+  fi
 fi
 
 
-nproc_per_node=2
+
 
 # 如果 nproc_per_node=1，那么就只使用第二块卡
 if [ ${nproc_per_node} -eq 1 ];then
